Move RQES types into their own package
CredentialId and CredentialID were both top level in
eu.europa.ec.eudi.verifier.endpoint.domain, so they compiled to the same
path on a case-insensitive filesystem and one overwrote the other, which
broke the build on macOS and Windows.

Give the RQES protocol constants and models their own
eu.europa.ec.eudi.verifier.endpoint.domain.rqes package, in the main and
the test sources, leaving the OpenId4VP base types where they were. The
two names no longer share a namespace.
