+++ src/main/kotlin/eu/europa/ec/eudi/verifier/endpoint/domain/rqes/RQES.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.verifier.endpoint.domain
+package eu.europa.ec.eudi.verifier.endpoint.domain.rqes
 
 /**
  * Remote Qualified Electronic Signature
